@@ -1,21 +1,22 @@
 module github.com/m-mizutani/zlog
 
-go 1.17
+go 1.18
 
 require (
 	github.com/k0kubun/colorstring v0.0.0-20150214042306-9440f1994b88
-	github.com/k0kubun/pp v3.0.1+incompatible
-	github.com/m-mizutani/goerr v0.1.2
+	github.com/k0kubun/pp/v3 v3.1.0
+	github.com/m-mizutani/goerr v0.1.6
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.7.0
+	github.com/stretchr/testify v1.7.1
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/google/uuid v1.3.0 // indirect
-	github.com/mattn/go-colorable v0.1.11 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/sys v0.0.0-20210927094055-39ccf1dd6fa6 // indirect
+	golang.org/x/text v0.3.7 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
